restructured for better usage for cloud computing
