--- res/ModelB25/conf.json
@@ -1,1 +1,0 @@
-{"modelA": "ModelA", "modelB": "ModelB", "model": "ModelB", "res_dir": "res", "param_path": "conf.json", "pl_size": 1, "pl_leader_reset_a": 0, "reset_max_u": 0.25, "pl_leader_tau": 0.5, "exact": "exact", "euler": "euler", "method": "exact", "timegap": 1.25, "dyn_coeff": 0.5, "reward_ev_coeff": 1, "reward_u_coeff": 0.1, "max_ep": 10, "max_ev": 10, "reset_ep_max": 3.5, "reset_max_ev": 3, "reset_max_a": 0.1, "action_high": 4.5, "action_low": -4.5, "re_scalar": 0.1, "normal": "normal", "uniform": "uniform", "rand_gen": "normal", "total_time_steps": 1000000, "sample_rate": 0.1, "episode_sim_time": 60, "steps_per_episode": 600, "number_of_episodes": 1666, "gamma": 0.99, "critic_lr": 0.001, "actor_lr": 0.0001, "std_dev": 0.02, "tau": 0.001, "batch_size": 64, "buffer_size": 100000, "show_env": false, "actor_fname": "actor.h5", "actor_picname": "actor.png", "actor_weights": "actor_weights.h5", "critic_fname": "critic.h5", "critic_picname": "critic.png", "critic_weights": "critic_weights.h5", "t_actor_fname": "target_actor.h5", "t_actor_picname": "target_actor.png", "t_actor_weights": "target_actor_weights.h5", "t_critic_fname": "target_critic.h5", "t_critic_picname": "target_critic.png", "t_critic_weights": "target_critic_weights.h5", "fig_path": "reward_curve.png", "zerofig_name": "zero", "constfig_name": "constant", "stepfig_name": "step", "rampfig_name": "ramp", "dirs": ["res"]}
