--- setup.py
@@ -1,10 +1,0 @@
-from distutils.core import setup
-
-setup(name='avddpg',
-      version='0.1',
-      description='DDPG implentation using tensorflow',
-      author='Christian Boin',
-      author_email='gward@python.net',
-      url='https://www.python.org/sigs/distutils-sig/',
-      packages=[],
-     )
