+++ src/setup.py
@@ -1,0 +1,10 @@
+from setuptools import setup
+
+setup(name='avddpg',
+      version='0.3',
+      description='DDPG implentation using tensorflow',
+      author='Christian Boin',
+      author_email='',
+      url='',
+      packages=[]
+     )
