+++ src/src/config.py
@@ -56,7 +56,7 @@ class Config():
         self.uniform = 'uniform'
         self.rand_gen = self.normal # which type of random numbers to use.
 
-        self.total_time_steps = 1000000
+        self.total_time_steps = 1000
 
         self.sample_rate = 0.1
         self.episode_sim_time = 60 # simulation time for a training episode
